Correct definition of RoleType Entity in party.md
Following the feedback from the mentor, this change addresses The definition of the RoleType Entity. The previous definition was "There is an Entity RoleType which stores the types of Roles available in the data model, we can also add a new role as per the requirement." This update ensures that the documentation accurately reflects The RoleType entity stores roles available OOTB like carrier, bill-to customer, ship-from vendor, and employee.