Changed how I think about the process.
The OrderFulfillmentWave is about starting to roll shipment for order.

For each ShipGroup,
Create Shipment
Then create picklist so picker can go collect items for packing all the shipments in the wave.

We want to think about picking in the context of Shipment workflow and not in the context of Order workflow.

We want to not put picklist information on OrderItem document in SOLR/SearchEngine. The Fulfillment status update on OrderItem solr document is all we need.